+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unpublished
+### 🚀 Features
+- Andes Tooltip fixed width option | Author: [@vveltri](https://github.com/vveltri)
+
 # 3.27.16
 ### 🚀 Features
 - Internal test (build public private (bintray & nexus) with fury create-version)